+++ src/main/java/com/forgestove/create_cyber_goggles/CreateCyberGoggles.java
@@ -1,18 +1,16 @@
 package com.forgestove.create_cyber_goggles;
 import com.forgestove.create_cyber_goggles.config.ModConfig;
+import com.forgestove.create_cyber_goggles.event.*;
 import me.shedaniel.autoconfig.AutoConfig;
 import me.shedaniel.autoconfig.serializer.Toml4jConfigSerializer;
-import net.minecraftforge.client.ConfigScreenHandler.ConfigScreenFactory;
-import net.minecraftforge.fml.ModLoadingContext;
-import net.minecraftforge.fml.common.Mod;
-@Mod(CreateCyberGoggles.ID)
-public class CreateCyberGoggles {
+import net.fabricmc.api.ClientModInitializer;
+public class CreateCyberGoggles implements ClientModInitializer {
 	public static final String ID = "create_cyber_goggles";
 	public static ModConfig config = AutoConfig.register(ModConfig.class, Toml4jConfigSerializer::new).getConfig();
-	public CreateCyberGoggles() {
-		ModLoadingContext.get().registerExtensionPoint(
-				ConfigScreenFactory.class,
-				() -> new ConfigScreenFactory((minecraft, screen) -> AutoConfig.getConfigScreen(ModConfig.class, screen).get())
-		);
+	@Override
+	public void onInitializeClient() {
+		KeyInput.register();
+		ModKeyMapping.register();
+		OverlayRenderer.register();
 	}
 }