+++ src/main/java/com/forgestove/create_cyber_goggles/event/KeyInput.java
@@ -1,13 +1,12 @@
 package com.forgestove.create_cyber_goggles.event;
 import com.forgestove.create_cyber_goggles.CreateCyberGoggles;
 import com.forgestove.create_cyber_goggles.config.ModConfig;
-import com.simibubi.create.AllMenuTypes;
 import com.simibubi.create.content.logistics.filter.*;
-import com.simibubi.create.content.logistics.stockTicker.*;
 import com.simibubi.create.foundation.blockEntity.SmartBlockEntity;
 import com.simibubi.create.foundation.blockEntity.behaviour.filtering.FilteringBehaviour;
+import com.simibubi.create.foundation.gui.ScreenOpener;
 import me.shedaniel.autoconfig.AutoConfig;
-import net.createmod.catnip.gui.ScreenOpener;
+import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
@@ -15,57 +14,38 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.HitResult.Type;
-import org.jetbrains.annotations.NotNull;
 
 import java.util.Collections;
 public class KeyInput {
-	public static StockTickerBlockEntity lastBlockEntity = null;
+	public static void register() {
+		ClientTickEvents.END_CLIENT_TICK.register(client -> {
+			KeyInput.toggleDiving();
+			KeyInput.openConfigScreen();
+			KeyInput.previewFilterScreen();
+		});
+	}
 	public static void toggleDiving() {
-		if (!ModKeyMapping.TOGGLE_DIVING.get().isDown()) return;
+		if (!ModKeyMapping.toggleDiving.isDown()) return;
 		var mc = Minecraft.getInstance();
 		var player = mc.player;
 		if (player == null || mc.screen != null) return;
 		var enabled = CreateCyberGoggles.config.armor.removeDivingBootsAffect;
 		CreateCyberGoggles.config.armor.removeDivingBootsAffect = !enabled;
-		player.displayClientMessage(
-				Component.translatable("message.%s.%s".formatted(
-						CreateCyberGoggles.ID,
-						enabled ? "enableDivingAffect" : "disableDivingAffect"
-				)), true
-		);
+		var component = Component.translatable("message.%s.%sableDivingAffect".formatted(CreateCyberGoggles.ID, enabled ? "en" : "dis"));
+		player.displayClientMessage(component, true);
 	}
 	public static void openConfigScreen() {
-		if (!ModKeyMapping.OPEN_CONFIG.get().isDown()) return;
+		if (!ModKeyMapping.openConfig.isDown()) return;
 		var mc = Minecraft.getInstance();
 		if (mc.screen != null) return;
 		mc.setScreen(AutoConfig.getConfigScreen(ModConfig.class, null).get());
 	}
-	public static void openStockScreen() {
-		if (!ModKeyMapping.OPEN_STOCK.get().isDown()) return;
-		var mc = Minecraft.getInstance();
-		if (mc.screen != null) return;
-		var player = mc.player;
-		if (player == null) return;
-		if (mc.hitResult == null) return;
-		if (mc.hitResult instanceof BlockHitResult blockHitResult && (
-				lastBlockEntity == null || blockHitResult.getType() == Type.BLOCK
-		)) {
-			if (mc.level == null) return;
-			if ((mc.level.getBlockEntity(blockHitResult.getBlockPos()) instanceof StockTickerBlockEntity stockTickerBlockEntity))
-				lastBlockEntity = stockTickerBlockEntity;
-		}
-		if (lastBlockEntity == null) return;
-		var type = AllMenuTypes.STOCK_KEEPER_REQUEST.get();
-		var inv = player.getInventory();
-		var menu = new StockKeeperRequestMenu(type, -1, inv, lastBlockEntity);
-		mc.setScreen(new StockKeeperRequestScreen(menu, inv, lastBlockEntity.getBlockState().getBlock().getName()));
-	}
-	public static void openFilterScreen() {
-		if (!ModKeyMapping.PREVIEW_FILTER.get().isDown()) return;
+	public static void previewFilterScreen() {
+		if (!ModKeyMapping.previewFilter.isDown()) return;
 		var mc = Minecraft.getInstance();
 		if (mc.screen != null) {
 			if (!(mc.screen instanceof AbstractContainerScreen<?> screen)) return;
-			var slot = screen.getSlotUnderMouse();
+			var slot = screen.hoveredSlot;
 			if (slot == null) return;
 			var item = slot.getItem();
 			if (!(item.getItem() instanceof FilterItem)) return;
@@ -83,7 +63,7 @@ public class KeyInput {
 			setFilterScreen(item);
 		}
 	}
-	public static void setFilterScreen(@NotNull ItemStack filter) {
+	public static void setFilterScreen(ItemStack filter) {
 		try {
 			var field = FilterItem.class.getDeclaredField("type");
 			field.setAccessible(true);
@@ -98,7 +78,6 @@ public class KeyInput {
 			switch (((Enum<?>) field.get(filterItem)).ordinal()) {
 				case 0 -> screen = new FilterScreen(FilterMenu.create(-1, inv, filter), inv, name);
 				case 1 -> screen = new AttributeFilterScreen(AttributeFilterMenu.create(-1, inv, filter), inv, name);
-				case 2 -> screen = new PackageFilterScreen(PackageFilterMenu.create(-1, inv, filter), inv, name);
 				default -> {
 					return;
 				}