+++ src/main/java/com/forgestove/create_cyber_goggles/event/OverlayRenderer.java
@@ -1,21 +1,18 @@
-package com.forgestove.create_cyber_goggles.render;
+package com.forgestove.create_cyber_goggles.event;
 import com.forgestove.create_cyber_goggles.CreateCyberGoggles;
 import com.forgestove.create_cyber_goggles.util.Util;
 import com.simibubi.create.content.kinetics.base.*;
 import com.simibubi.create.content.logistics.depot.DepotBlockEntity;
-import com.simibubi.create.content.logistics.packager.PackagerBlockEntity;
-import net.createmod.catnip.math.VecHelper;
+import com.simibubi.create.foundation.utility.VecHelper;
+import net.fabricmc.fabric.api.client.rendering.v1.HudRenderCallback;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.world.phys.BlockHitResult;
-import net.minecraftforge.client.event.RegisterGuiOverlaysEvent;
-import net.minecraftforge.client.gui.overlay.*;
-import org.jetbrains.annotations.NotNull;
 public class OverlayRenderer {
-	public static void register(@NotNull RegisterGuiOverlaysEvent event) {
-		event.registerAbove(VanillaGuiOverlay.HOTBAR.id(), "goggle_overlay", OverlayRenderer::renderOverlay);
+	public static void register() {
+		HudRenderCallback.EVENT.register(OverlayRenderer::renderOverlay);
 	}
-	public static void renderOverlay(ForgeGui forgeGui, GuiGraphics guiGraphics, float v, int i, int i1) {
+	public static void renderOverlay(GuiGraphics guiGraphics, float tickDelta) {
 		var mc = Minecraft.getInstance();
 		var player = mc.player;
 		if (player == null || mc.isPaused() || mc.screen != null) return;
@@ -25,8 +22,6 @@ public class OverlayRenderer {
 		var renderExtraItems = CreateCyberGoggles.config.goggles.renderExtraItems;
 		if (renderExtraItems && blockEntity instanceof DepotBlockEntity depotBlockEntity)
 			Util.renderItemStack(guiGraphics, depotBlockEntity.getHeldItem());
-		else if (renderExtraItems && blockEntity instanceof PackagerBlockEntity packagerBlockEntity)
-			Util.renderItemStack(guiGraphics, packagerBlockEntity.heldBox);
 		else if (CreateCyberGoggles.config.goggles.enableKineticEffect && blockEntity instanceof KineticBlockEntity kineticBlockEntity) {
 			if (!blockHitResult.getBlockPos().equals(kineticBlockEntity.getBlockPos())) return;
 			var speed = kineticBlockEntity.getSpeed();